add a signature test and forge fmt

+++ src/Wallet.sol
@@ -3,7 +3,9 @@ pragma solidity ^0.8.21;
 
 contract Wallet {
     address payable public owner;
+
     event Deposit(address account, uint256 amount);
+
     constructor() payable {
         owner = payable(msg.sender);
     }
@@ -12,17 +14,17 @@ contract Wallet {
         emit Deposit(msg.sender, msg.value);
     }
 
-    function withdraw(uint _amount) external {
+    function withdraw(uint256 _amount) external {
         require(msg.sender == owner, "caller is not owner");
         payable(msg.sender).transfer(_amount);
     }
 
     function setOwner(address _owner) external {
-      require(msg.sender == owner, "caller is not owner");
-      owner = payable(_owner);
+        require(msg.sender == owner, "caller is not owner");
+        owner = payable(_owner);
     }
 
-    function getBalance() external view returns (uint) {
+    function getBalance() external view returns (uint256) {
         return address(this).balance;
     }
 }
